Handle NGS multi-disc-releases; fixes #1
Pre-NGS there was one release object per disc. Discs where grouped in
release groups. Now with NGS there are
multiple "media" in one release object so the MB webservice 1 we use
gives us the tracks of all discs.
An offset is used to set the position of the disc in the given
tracklist. For the first and last disc this can get calculated because
we have the overall number of tracks in the release and the number of
tracks on our disc. For "middle disc" the user needs to give us the
offset.
The command help was updated to inform the user.