multi-disc check ONLY on track count mismatch, re #1
When the number of tracks reported on the disc and the number of tracks
given on the release match, we do not need to check the list of discIds.
This just means we have multiple discIds for a single disc.
There is only a problem when there is another "real" disc, but then the
track counts would not match.