multiple backend support; cdda2wav possible again (close #10)
Many changes in the code to make multiple backends possible.
cdda2wav is added as proof of concept.
The actual gathering is completely the same as with icedax.

We check if one of the supported backends is available and
if it is only a symlink to another backend.
The priority is now (highest first): cdda2wav, icedax