finally drutil/Mac OS X support, close #12
On Mac OS X the device has to be given as a positive number.
1 is your first cd drive (default), 2 the second..
The script will find out what the name of the corresponding device node
is.

We will probably never support Windows. There are other tools and
windows has a crappy (built in) commandline. Using cygwin would be overkill.