find python2 with a shellscript
This python, python2 and python3 thing is still difficult.
Not every distribution has a python2 link and just "python" is only
python version 2 for half of the distributions, too.
Added a shellscript that automatically executes the script with
python2.7, 2.6 or 2.5. This will run with 2.4, but then ctypes needs to
be added. And python 2.4 is ancient.

For users distributions including python2 as a symlink this shellscript is
completely useless. The same for users that are not afraid of changing
the first line of the file ;-)