handle bonus DVDs gracefully, re #16
Having a track count mismatch between the disc and the release given by
the (version 1) API with just one disc ID is a weird case.
There either is a real problem with some data tracks on a disc,
or there is another disc in this release without a disc ID.

This happens when a release with just 1 CD has an additional bonus DVD
that is also tracked in MB. DVDs don't have disc IDs.

We handle the latter case gracefully now and guess the track offset as
0.