Don't assign ISRCs to empty track numbers (cdrdao).
If you have a release that stores ISRCs both on their own and in the
CD-TEXT, the cdrdao parser would assign the ISRC both to the
trackNumber ("[1, ISRC]"), but also to an empty track ("[None, ISRC]").
This would cause errors on line 658[1] as you cannot add type None
to type Int.

[1] "track = tracks[trackNumber + trackOffset - 1]"