create Disc object and use it
Due to the creation of getDisc we had to work with global
variables in the function and it got messy.
The previous commit fixed a bug, but not the mess with globals.

We should have only a Disc object as global and take the id and
trackCount from there.

Other globals are not touched and might need additional objects.