ignore stubs, fix #23
A search with a discId can yield a stub as a release.
These don't have musicbrainz ids and certainly don't have ISRCs
attached. We just ignore them and let the user submit the disc as a new
release. The submission process in the browser will present another
chance to use the stub data.