add libcdio/cd-info backend, close #22
cd-info is part of libcdio and prints a lot of information about audio
discs. We suppress cddb and audio information.
This tool is cleaner than cdda2wav, but cdrdao is still the first
choice. I don't know if cd-info also fetches ISRCs from CD-TEXT.
The quality of the ISRC gathering seems to be identical to the other
tools.
In theory this is supposed to work on windows, but there is no build
available for download.