check if "which" is sane, close #30
Windows unxutils includes which 2.4 which is broken in terms of not
including the current directory and needing .exe appended.
We don't check for the version,
but if "which which" actually works (on all platforms).

On Windows we suggest updating to GnuWin32 version of which,
when a which is found that is not sane.