pep8: fix some (not all) methods names, close #49
Some classes will be used in isrcsubmit2 and I did change
the function names in some.
However, some classes are near direct wrappers for
python-musicbrainz2 classes or use these functions/methods.
I didn't change the style for these to be consistent.

The code is still far from pep8 "compliant",
but a step in the right direction.