always use drive numbers as options.device on Mac, close #55
We recently found out that using numbers for drives on Mac
is the way to go anyways. see: PICARD-123

So we always use that and only convert to raw device
right before using libdiscid/discisrc on Mac/Darwin.

Previously we had a bit of a hackish way to change between these and
that broke.