be less explicit on browser option
We now have several ways trying to detect the default browser:
xdg-open, x-www-browser, webbrowser module and a static list.
Explaining all of this in the description makes it quite difficult.

Most people only want to know that they can, but don't have to set this
option in order to get a browser "connected".

The few users that actually do care will find the BROWSERS variable
right at the top of the script.