Fix the package declaration
Since isrcsubmit.py is currently contained in a single file, that means
package declaration is a bit different. Instead of specifying a
`packages` entry, we just mark the singular module file in `py_modules`
instead.

Note that the script still isn’t working, but that requires a bit more
refactoring, so that’ll be done in a separate commit.

Follow-up to 73c820da82701e560b51e3a0d29128bffcc28528